# web/src/api/openapi.json
{
  "components": {
    "schemas": {
...
      "FilterLogic": {
        "description": "How multiple resource-presence filters combine.",
        "enum": [
          "and",
          "or"
        ],
        "title": "FilterLogic",
        "type": "string"
      },
...
      "MetricFilter": {
        "enum": [
          "assemblies",
          "annotations",
          "rna_seq",
          "long_read_rna_seq"
        ],
        "title": "MetricFilter",
        "type": "string"
      },
...
      "SortOrder": {
        "enum": [
          "asc",
          "desc"
        ],
        "title": "SortOrder",
        "type": "string"
      },
      "TargetRank": {
        "enum": [
          "kingdom",
          "phylum",
          "class",
          "order",
          "family",
          "genus",
          "species"
        ],
        "title": "TargetRank",
        "type": "string"
      },
...
      "TaxonPage": {
        "properties": {
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "next": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next"
          },
          "previous": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Previous"
          },
          "results": {
            "items": {
              "$ref": "#/components/schemas/Taxon"
            },
            "title": "Results",
            "type": "array"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "total",
          "limit",
          "next",
          "previous",
          "results"
        ],
        "title": "TaxonPage",
        "type": "object"
      },
...
      "TaxonSort": {
        "enum": [
          "species",
          "resources.assemblies.covered",
          "resources.assemblies.missing",
          "resources.assemblies.total",
          "resources.annotations.covered",
          "resources.annotations.missing",
          "resources.annotations.total",
          "resources.rna_seq.covered",
          "resources.rna_seq.missing",
          "resources.rna_seq.total",
          "resources.long_read_rna_seq.covered",
          "resources.long_read_rna_seq.missing",
          "resources.long_read_rna_seq.total",
          "composition.complete",
          "composition.chromosome",
          "composition.scaffold",
          "composition.contig",
          "composition.reference",
          "name"
        ],
        "title": "TaxonSort",
        "type": "string"
      },
...
    }
  },
  "info": {
    "title": "EukaHub API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/taxons": {
      "get": {
        "description": "Taxa with their counts: a name search (``q``), a taxon's children\n(``parent``), every taxon of a rank under a taxon (``within`` and ``rank``), or\nchosen taxa (``taxids``), narrowed by the data they have. Sort by\n``resources.<resource>.missing`` for the groups with the most species still\nmissing it. Their quality stats are in ``/taxons/stats``.",
        "operationId": "taxons_taxons_get",
        "parameters": [
          {
            "description": "The path of a number in a taxon: `species`, `resources.<resource>.covered`, `.missing` or `.total`, or `composition.<field>`; or `name`. Without it: relevance for `q`, else `species`.",
            "in": "query",
            "name": "sort_by",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/TaxonSort"
                },
                {
                  "type": "null"
                }
              ],
              "description": "The path of a number in a taxon: `species`, `resources.<resource>.covered`, `.missing` or `.total`, or `composition.<field>`; or `name`. Without it: relevance for `q`, else `species`.",
              "title": "Sort By"
            }
          },
          {
            "in": "query",
            "name": "sort_order",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/SortOrder",
              "default": "desc"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 25,
              "maximum": 1000,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "description": "``next`` or ``previous`` from a page with the same sort, for the page after or before it.",
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 1000,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "``next`` or ``previous`` from a page with the same sort, for the page after or before it.",
              "title": "Cursor"
            }
          },
          {
            "description": "Only taxa whose name contains this text, ignoring case (or is spelled like it, with `fuzzy`). The root and 'cellular organisms' are left out.",
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 100,
                  "minLength": 3,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Only taxa whose name contains this text, ignoring case (or is spelled like it, with `fuzzy`). The root and 'cellular organisms' are left out.",
              "title": "Q"
            }
          },
          {
            "description": "Match `q` by spelling instead, for a misspelt name.",
            "in": "query",
            "name": "fuzzy",
            "required": false,
            "schema": {
              "default": false,
              "description": "Match `q` by spelling instead, for a misspelt name.",
              "title": "Fuzzy",
              "type": "boolean"
            }
          },
          {
            "description": "Only the direct children of this taxon.",
            "in": "query",
            "name": "parent",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Only the direct children of this taxon.",
              "title": "Parent"
            }
          },
          {
            "description": "Only rows on this taxon or below it (e.g. 40674 for mammals).",
            "in": "query",
            "name": "within",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Only rows on this taxon or below it (e.g. 40674 for mammals).",
              "title": "Within"
            }
          },
          {
            "description": "Only taxa of this rank.",
            "in": "query",
            "name": "rank",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/TargetRank"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Only taxa of this rank.",
              "title": "Rank"
            }
          },
          {
            "description": "Only these taxa: comma-separated taxids, at most 100.",
            "in": "query",
            "name": "taxids",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Only these taxa: comma-separated taxids, at most 100.",
              "title": "Taxids"
            }
          },
          {
            "description": "Only taxa with data for these resources (their names in /config).",
            "in": "query",
            "name": "filter",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "$ref": "#/components/schemas/MetricFilter"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Only taxa with data for these resources (their names in /config).",
              "title": "Filter"
            }
          },
          {
            "description": "Whether `filter` needs every resource (`and`) or any (`or`).",
            "in": "query",
            "name": "logic",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/FilterLogic",
              "default": "and",
              "description": "Whether `filter` needs every resource (`and`) or any (`or`)."
            }
          },
          {
            "description": "Only taxa with data for at least one resource.",
            "in": "query",
            "name": "exclude_empty",
            "required": false,
            "schema": {
              "default": false,
              "description": "Only taxa with data for at least one resource.",
              "title": "Exclude Empty",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaxonPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "An error, as problem details (RFC 9457)."
          }
        },
        "summary": "Taxons",
        "tags": [
          "taxons"
        ]
      }
    },
    "/taxons/aggregates": {
      "get": {
        "description": "Species count, per-resource coverage and quality stats for a set of clades:\nthe clades in ``include`` minus the clades inside them in ``exclude`` (e.g.\nfish as Vertebrata minus Tetrapoda). A clade inside an excluded one can be\nincluded again. Counts are sums and differences of the clades' rollups;\nquality stats are computed from the records in the set. A list, like every\ncollection, holding the one set asked for.",
        "operationId": "aggregate_taxons_aggregates_get",
        "parameters": [
          {
            "description": "Comma-separated taxids of the clades to add up (1-20, e.g. 7742).",
            "in": "query",
            "name": "include",
            "required": true,
            "schema": {
              "description": "Comma-separated taxids of the clades to add up (1-20, e.g. 7742).",
              "title": "Include",
              "type": "string"
            }
          },
          {
            "description": "Comma-separated taxids of clades inside them to leave out (e.g. 32523).",
            "in": "query",
            "name": "exclude",
            "required": false,
            "schema": {
              "default": "",
              "description": "Comma-separated taxids of clades inside them to leave out (e.g. 32523).",
              "title": "Exclude",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AggregatePage"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "An error, as problem details (RFC 9457)."
          }
        },
        "summary": "Aggregate",
        "tags": [
          "taxons"
        ]
      }
    },
    "/taxons/report": {
      "get": {
        "description": "Every taxon ``/taxons`` lists for the same filters and sort, as a streamed\nTSV download: taxid, name, species count, then the species with each resource\nand the total of each resource. Not cached: a report runs to tens of MB, and\nnginx's cache is kept for the small responses the interface repeats.",
        "operationId": "taxons_report_taxons_report_get",
        "parameters": [
          {
            "description": "The path of a number in a taxon: `species`, `resources.<resource>.covered`, `.missing` or `.total`, or `composition.<field>`; or `name`. Without it: relevance for `q`, else `species`.",
            "in": "query",
            "name": "sort_by",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/TaxonSort"
                },
                {
                  "type": "null"
                }
              ],
              "description": "The path of a number in a taxon: `species`, `resources.<resource>.covered`, `.missing` or `.total`, or `composition.<field>`; or `name`. Without it: relevance for `q`, else `species`.",
              "title": "Sort By"
            }
          },
          {
            "in": "query",
            "name": "sort_order",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/SortOrder",
              "default": "desc"
            }
          },
          {
            "description": "Only taxa whose name contains this text, ignoring case (or is spelled like it, with `fuzzy`). The root and 'cellular organisms' are left out.",
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 100,
                  "minLength": 3,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Only taxa whose name contains this text, ignoring case (or is spelled like it, with `fuzzy`). The root and 'cellular organisms' are left out.",
              "title": "Q"
            }
          },
          {
            "description": "Match `q` by spelling instead, for a misspelt name.",
            "in": "query",
            "name": "fuzzy",
            "required": false,
            "schema": {
              "default": false,
              "description": "Match `q` by spelling instead, for a misspelt name.",
              "title": "Fuzzy",
              "type": "boolean"
            }
          },
          {
            "description": "Only the direct children of this taxon.",
            "in": "query",
            "name": "parent",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Only the direct children of this taxon.",
              "title": "Parent"
            }
          },
          {
            "description": "Only rows on this taxon or below it (e.g. 40674 for mammals).",
            "in": "query",
            "name": "within",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Only rows on this taxon or below it (e.g. 40674 for mammals).",
              "title": "Within"
            }
          },
          {
            "description": "Only taxa of this rank.",
            "in": "query",
            "name": "rank",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/TargetRank"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Only taxa of this rank.",
              "title": "Rank"
            }
          },
          {
            "description": "Only these taxa: comma-separated taxids, at most 100.",
            "in": "query",
            "name": "taxids",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Only these taxa: comma-separated taxids, at most 100.",
              "title": "Taxids"
            }
          },
          {
            "description": "Only taxa with data for these resources (their names in /config).",
            "in": "query",
            "name": "filter",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "$ref": "#/components/schemas/MetricFilter"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Only taxa with data for these resources (their names in /config).",
              "title": "Filter"
            }
          },
          {
            "description": "Whether `filter` needs every resource (`and`) or any (`or`).",
            "in": "query",
            "name": "logic",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/FilterLogic",
              "default": "and",
              "description": "Whether `filter` needs every resource (`and`) or any (`or`)."
            }
          },
          {
            "description": "Only taxa with data for at least one resource.",
            "in": "query",
            "name": "exclude_empty",
            "required": false,
            "schema": {
              "default": false,
              "description": "Only taxa with data for at least one resource.",
              "title": "Exclude Empty",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "An error, as problem details (RFC 9457)."
          }
        },
        "summary": "Taxons Report",
        "tags": [
          "taxons"
        ]
      }
    },
...
    "/taxons/{taxid}": {
      "get": {
        "description": "One taxon, the same object ``/taxons`` lists: species count, per-resource\ncoverage and assembly composition. A species, an informal species or a finer\ntaxon also has ``direct``: its records attached to the taxon itself rather than\nto a finer taxon below it.",
        "operationId": "taxon_taxons__taxid__get",
        "parameters": [
          {
            "in": "path",
            "name": "taxid",
            "required": true,
            "schema": {
              "title": "Taxid",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Taxon"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "An error, as problem details (RFC 9457)."
          }
        },
        "summary": "Taxon",
        "tags": [
          "taxons"
        ]
      }
    },
    "/taxons/{taxid}/ancestors": {
      "get": {
        "description": "The root, every taxon below it down to this one, and this one, in that order,\nas the same objects ``/taxons`` lists. The whole lineage, unpaged: it is at\nmost a few dozen taxa.",
        "operationId": "taxon_ancestors_taxons__taxid__ancestors_get",
        "parameters": [
          {
            "in": "path",
            "name": "taxid",
            "required": true,
            "schema": {
              "title": "Taxid",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Taxon"
                  },
                  "title": "Response Taxon Ancestors Taxons  Taxid  Ancestors Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "An error, as problem details (RFC 9457)."
          }
        },
        "summary": "Taxon Ancestors",
        "tags": [
          "taxons"
        ]
      }
    },
    "/taxons/{taxid}/stats": {
      "get": {
        "description": "The quality stats (best BUSCO, median genes, genome size and N50) of the\nrecords on or below one taxon, the same object ``/taxons/stats`` lists,\ncomputed at build time.",
        "operationId": "taxon_stats_taxons__taxid__stats_get",
        "parameters": [
          {
            "in": "path",
            "name": "taxid",
            "required": true,
            "schema": {
              "title": "Taxid",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaxonStats"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "An error, as problem details (RFC 9457)."
          }
        },
        "summary": "Taxon Stats",
        "tags": [
          "taxons"
        ]
      }
    }
  }
}